+++ HISTORY.md
@@ -5,6 +5,7 @@ unreleased
   * Fully URL-encode the pathname in the 404 message
   * Only include the pathname in the 404 message
   * Send complete HTML document
+  * Set `Content-Security-Policy: default-src 'self'` header
   * deps: debug@2.6.1
     - Allow colors in workers
     - Deprecated `DEBUG_FD` environment variable set to `3` or higher